V : 0.10.1

+++ sources/delete.rb
@@ -14,37 +14,48 @@ def del(path, extension)
   end
 end
 
-def erase_files(db, name)
-  data = db.get_manga(name)
-  params = Params.instance.get_params
-  path_site = ''
-  if data[3] == 'http://mangafox.me/'
-    path_site = 'mangafox/'
-  else
-    puts 'unmanaged site for deletion : ' + data[3]
-    puts 'please report this error'
-    exit 4
-  end
+def files_to_delete(params, path_site, name)
   del(params[1] + path_site + 'html/' + name, '.html')
   del(params[1] + path_site + 'mangas/' + name, '.jpg')
 end
 
 def confirm_delete (db, name, delete_files)
-    puts "going to delete #{name} from database " + ((delete_files) ? 'and delete files' : 'only')
-    puts "Write 'YES' to continue"
-    ret = STDIN.gets.chomp
-    puts ''
-    if ret == 'YES'
-      if delete_files
-        puts 'deleting files'
-        erase_files(db, name)
-      end
-      db.delete_manga(name)
-      puts "deleted #{name} from database"
-      return true
+  unless db.manga_in_data?(name)
+    puts "did not find #{name} in database"
+    return false
+  end
+  puts "going to delete #{name} from database " + ((delete_files) ? 'and delete files' : 'only')
+  puts "Write 'YES' to continue"
+  ret = STDIN.gets.chomp
+  puts ''
+  if ret == 'YES'
+    if delete_files
+      puts 'deleting files'
+      erase_files(db, name)
     end
-    puts "did not delete #{name}"
-    false
+    db.delete_manga(name)
+    puts "deleted #{name} from database"
+    return true
+  end
+  puts "did not delete #{name}"
+  false
+end
+
+def erase_files(db, name)
+  data = db.get_manga(name)
+  params = Params.instance.get_params
+  if data != nil
+    if data[3] == 'http://mangafox.me/'
+      files_to_delete(params, 'mangafox/', name)
+    else
+      puts 'unmanaged site for deletion : ' + data[3]
+      puts 'please report this error'
+      exit 4
+    end
+  else
+    puts 'did not find manga in database'
+    exit 1
+  end
 end
 
 def delete(db, delete_files = false)
